Add `InputType` and move option types to their own file
